requirements.txt
streamlit>=1.35.0
pillow
numpy
streamlit-drawable-canvas==0.9.3